Añadida carpeta con shaders y refactorizacion de RenderQuad para que sea usable. Corregido bug en el que los sistemas cogen a una entidad mas de una vez. Creada clase BluePhoenix que engloba todo el proceso del game engine

+++ BluePhoenix/src/Components/Transform.h
@@ -3,6 +3,6 @@
 
 struct Transform
 {
-	double x = 0, y = 0;
+	float x = 0, y = 0;
 };
